requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
requests>=2.31.0
redis>=5.0.1
rq>=1.15.1
google-generativeai>=0.3.2
python-dotenv>=1.0.1
gunicorn>=21.2.0
